+++ CHANGELOG.md
@@ -9,6 +9,9 @@ This is the version currently published on master.
 + Added **Species** endpoints
   + GET Single
   + GET List
++ Added **Vehicles** endpoints
+  + GET Single
+  + GET List
 
 ### [Under Development] - Changed